Created Ship class
